@@ -17,28 +17,27 @@
     </description>
     <inceptionYear>2008</inceptionYear>
 
-    <prerequisites>
-        <maven>3.0</maven>
-    </prerequisites>
-
     <!-- ================= Default Properties ================= -->
     <properties>
         <restlet.version>2.4.3</restlet.version>
+        <rdf4j.version>4.2.2</rdf4j.version>
         <solr.version>9.0.0</solr.version>
-        <sesame.version>2.7.16</sesame.version>
         <slf4j.version>1.7.36</slf4j.version>
         <log4j.version>2.19.0</log4j.version>
         <servlet.version>3.1.0</servlet.version>
         <xerces.version>2.12.2</xerces.version>
         <jaxb.version>2.3.1</jaxb.version>
         <guava.version>31.1-jre</guava.version>
-        <jetty.version>9.4.49.v20220914</jetty.version>
-	<jsonorg.version>20220320</jsonorg.version>
+        <jetty.version>9.4.50.v20221201</jetty.version>
+        <jsonorg.version>20220924</jsonorg.version>
+        <jackson.version>2.14.1</jackson.version>
+        <jackson-databind.version>2.14.1</jackson-databind.version>
         <site.url>https://entrystore.org</site.url>
         <distribution.snapshot.url>https://entrystore.org</distribution.snapshot.url>
         <distribution.url>https://entrystore.org</distribution.url>
         <license.licenseName>apache_v2</license.licenseName>
-        <maven.compiler.release>11</maven.compiler.release>
+        <maven.compiler.release>17</maven.compiler.release>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
     <!-- ================= Developers ================= -->
