@@ -16,6 +16,25 @@
 
 package org.entrystore.rest.util;
 
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonGenerator;
+import org.eclipse.rdf4j.model.BNode;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Literal;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.Resource;
+import org.eclipse.rdf4j.model.Statement;
+import org.eclipse.rdf4j.model.Value;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.model.impl.LinkedHashModel;
+import org.eclipse.rdf4j.model.impl.SimpleValueFactory;
+import org.entrystore.repository.util.NS;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.IOException;
 import java.io.StringWriter;
 import java.util.HashMap;
@@ -23,23 +42,7 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
 
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonGenerationException;
-import com.fasterxml.jackson.core.JsonGenerator;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.openrdf.model.BNode;
-import org.openrdf.model.Graph;
-import org.openrdf.model.Literal;
-import org.openrdf.model.Resource;
-import org.openrdf.model.Statement;
-import org.openrdf.model.URI;
-import org.openrdf.model.Value;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.model.impl.GraphImpl;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import static org.eclipse.rdf4j.model.util.Values.iri;
 
 /**
  * A utility class to help converting Sesame Graphs from and to RDF/JSON.
@@ -48,7 +51,11 @@ import org.slf4j.LoggerFactory;
  */
 public class RDFJSON {
 	
-	private static Logger log = LoggerFactory.getLogger(RDFJSON.class);
+	private static final Logger log = LoggerFactory.getLogger(RDFJSON.class);
+
+	private static IRI dtString = iri(NS.xsd, "string");
+
+	private static IRI dtLangString = iri(NS.rdf, "langString");
 
 	/**
 	 * Implementation using the json.org API.
@@ -58,10 +65,10 @@ public class RDFJSON {
 	 *            Graph.
 	 * @return A Sesame Graph if successful, otherwise null.
 	 */
-	public static Graph rdfJsonToGraph(JSONObject json) {
-		Graph result = new GraphImpl();
-		HashMap<String, BNode> id2bnode = new HashMap<String, BNode>();
-		ValueFactory vf = result.getValueFactory();
+	public static Model rdfJsonToGraph(JSONObject json) {
+		Model result = new LinkedHashModel();
+		HashMap<String, BNode> id2bnode = new HashMap<>();
+		ValueFactory vf = SimpleValueFactory.getInstance();
 
 		try {
 			JSONObject input = json;
@@ -78,7 +85,7 @@ public class RDFJSON {
 							id2bnode.put(subjStr, (BNode) subject);
 						}
 					} else {
-						subject = vf.createURI(subjStr);
+						subject = vf.createIRI(subjStr);
 					}
 				} catch (IllegalArgumentException iae) {
 					subject = vf.createBNode();
@@ -87,7 +94,7 @@ public class RDFJSON {
 				Iterator<String> predicates = pObj.keys();
 				while (predicates.hasNext()) {
 					String predStr = predicates.next();
-					URI predicate = vf.createURI(predStr);
+					IRI predicate = vf.createIRI(predStr);
 					JSONArray predArr = pObj.getJSONArray(predStr);
 					for (int i = 0; i < predArr.length(); i++) {
 						Value object = null;
@@ -103,10 +110,13 @@ public class RDFJSON {
 						String lang = null;
 						if (obj.has("lang")) {
 							lang = obj.getString("lang");
+							if (lang.trim().isEmpty()) {
+								lang = null;
+							}
 						}
-						URI datatype = null;
+						IRI datatype = null;
 						if (obj.has("datatype")) {
-							datatype = vf.createURI(obj.getString("datatype"));
+							datatype = vf.createIRI(obj.getString("datatype"));
 						}
 						if ("literal".equals(type)) {
 							if (lang != null) {
@@ -124,7 +134,7 @@ public class RDFJSON {
 								id2bnode.put(value, (BNode) object);
 							}
 						} else if ("uri".equals(type)) {
-							object = vf.createURI(value);
+							object = vf.createIRI(value);
 						}
 						result.add(subject, predicate, object);
 					}
@@ -138,7 +148,7 @@ public class RDFJSON {
 		return result;
 	}
 	
-	public static Graph rdfJsonToGraph(String json) {
+	public static Model rdfJsonToGraph(String json) {
 		try {
 			return rdfJsonToGraph(new JSONObject(json));
 		} catch (JSONException e) {
@@ -150,37 +160,40 @@ public class RDFJSON {
 	private static JSONObject getValue(Value v) {
 		JSONObject valueObj = new JSONObject();
 		if (v instanceof BNode && !v.stringValue().startsWith("_:")) {
-			valueObj.put("value", "_:"+v.stringValue());
+			valueObj.put("value", "_:" + v.stringValue());
 		} else {
 			valueObj.put("value", v.stringValue());
 		}
-		if (v instanceof Literal) {
+		if (v instanceof Literal l) {
 			valueObj.put("type", "literal");
-			Literal l = (Literal) v;
-			if (l.getLanguage() != null) {
-				valueObj.put("lang", l.getLanguage());
+			if (l.getLanguage().isPresent()) {
+				valueObj.put("lang", l.getLanguage().get());
 			} else if (l.getDatatype() != null) {
-				valueObj.put("datatype", l.getDatatype().stringValue());
+				IRI dataType = l.getDatatype();
+				// we ignore data types for strings (as introduced by RDF 1.1) to be compatible with RDF 1.0
+				if (!dataType.equals(dtString) && !dataType.equals(dtLangString)) {
+					valueObj.put("datatype", dataType.stringValue());
+				}
 			}
 		} else if (v instanceof BNode) {
 			valueObj.put("type", "bnode");
-		} else if (v instanceof URI) {
+		} else if (v instanceof IRI) {
 			valueObj.put("type", "uri");
 		}
 		return valueObj;
 	}
 
-	public static JSONObject graphToRdfJsonObject(Graph graph) {
+	public static JSONObject graphToRdfJsonObject(Model graph) {
 		try {
 			//First build the json structure using maps to avoid iterating through the graph more than once
-			HashMap<Resource, HashMap<URI, JSONArray>> struct = new HashMap<Resource, HashMap<URI, JSONArray>>();
+			HashMap<Resource, HashMap<IRI, JSONArray>> struct = new HashMap<>();
 			for (Statement stmt : graph) {
 				Resource subject = stmt.getSubject();
-				URI predicate = stmt.getPredicate();
+				IRI predicate = stmt.getPredicate();
 				Value object = stmt.getObject();
-				HashMap<URI, JSONArray> pred2values = struct.get(subject);
+				HashMap<IRI, JSONArray> pred2values = struct.get(subject);
 				if (pred2values == null) {
-					pred2values = new HashMap<URI, JSONArray>();
+					pred2values = new HashMap<>();
 					struct.put(subject, pred2values);
 					JSONArray values = new JSONArray();
 					pred2values.put(predicate, values);
@@ -199,8 +212,8 @@ public class RDFJSON {
 			JSONObject result = new JSONObject(); //Top level object
 			for (Resource subject : struct.keySet()) {
 				JSONObject predicateObj = new JSONObject(); //Predicate object where each predicate is a key
-				HashMap<URI, JSONArray> pred2values = struct.get(subject);
-				for (URI predicate : pred2values.keySet()) {
+				HashMap<IRI, JSONArray> pred2values = struct.get(subject);
+				for (IRI predicate : pred2values.keySet()) {
 					predicateObj.put(predicate.stringValue(), pred2values.get(predicate)); //The value is an array of objects
 				}
 
@@ -224,7 +237,7 @@ public class RDFJSON {
 	 *            A Sesame Graph.
 	 * @return An RDF/JSON string if successful, otherwise null.
 	 */
-	public static String graphToRdfJson(Graph graph) {
+	public static String graphToRdfJson(Model graph) {
 		JSONObject obj = graphToRdfJsonObject(graph);
 		if (obj != null) {
 			return obj.toString(2);
@@ -240,7 +253,7 @@ public class RDFJSON {
 	 *            A Sesame Graph.
 	 * @return An RDF/JSON string if successful, otherwise null.
 	 */
-	public static String graphToRdfJsonJackson(Graph graph) {
+	public static String graphToRdfJsonJackson(Model graph) {
 		JsonFactory f = new JsonFactory();
 		StringWriter sw = new StringWriter();
 		JsonGenerator g = null;
@@ -264,33 +277,30 @@ public class RDFJSON {
 				} else {
 					g.writeObjectFieldStart(subject.stringValue()); // subject					
 				}
-				Set<URI> predicates = new HashSet<URI>();
-				Iterator<Statement> s2 = graph.match(subject, null, null);
-				while (s2.hasNext()) {
-					predicates.add(s2.next().getPredicate());
+				Set<IRI> predicates = new HashSet<>();
+				for (Statement statement : graph.filter(subject, null, null)) {
+					predicates.add(statement.getPredicate());
 				}
-				for (URI predicate : predicates) {
+				for (IRI predicate : predicates) {
 					g.writeArrayFieldStart(predicate.stringValue()); // predicate
-					Iterator<Statement> stmnts = graph.match(subject, predicate, null);
-					while (stmnts.hasNext()) {
-						Value v = stmnts.next().getObject();
+					for (Statement statement : graph.filter(subject, predicate, null)) {
+						Value v = statement.getObject();
 						g.writeStartObject(); // value
-						if (v instanceof BNode && ! v.stringValue().startsWith("_:")) {
-							g.writeStringField("value", "_:"+v.stringValue());							
+						if (v instanceof BNode && !v.stringValue().startsWith("_:")) {
+							g.writeStringField("value", "_:" + v.stringValue());
 						} else {
-							g.writeStringField("value", v.stringValue());							
+							g.writeStringField("value", v.stringValue());
 						}
-						if (v instanceof Literal) {
+						if (v instanceof Literal l) {
 							g.writeStringField("type", "literal");
-							Literal l = (Literal) v;
-							if (l.getLanguage() != null) {
-								g.writeStringField("lang", l.getLanguage());
+							if (l.getLanguage().isPresent()) {
+								g.writeStringField("lang", l.getLanguage().get());
 							} else if (l.getDatatype() != null) {
 								g.writeStringField("datatype", l.getDatatype().stringValue());
 							}
 						} else if (v instanceof BNode) {
 							g.writeStringField("type", "bnode");
-						} else if (v instanceof URI) {
+						} else if (v instanceof IRI) {
 							g.writeStringField("type", "uri");
 						}
 						g.writeEndObject(); // value
@@ -302,10 +312,8 @@ public class RDFJSON {
 			g.writeEndObject(); // root object
 			g.close();
 			return sw.toString();
-		} catch (JsonGenerationException e) {
+		} catch (IOException e) {
 			log.error(e.getMessage(), e);
-		} catch (IOException ioe) {
-			log.error(ioe.getMessage(), ioe);
 		}
 		return null;
 	}
