@@ -17,6 +17,15 @@
 
 package org.entrystore.impl;
 
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Literal;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.Resource;
+import org.eclipse.rdf4j.model.Statement;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.repository.RepositoryConnection;
+import org.eclipse.rdf4j.repository.RepositoryException;
+import org.eclipse.rdf4j.repository.RepositoryResult;
 import org.entrystore.AuthorizationException;
 import org.entrystore.Context;
 import org.entrystore.Data;
@@ -35,15 +44,6 @@ import org.entrystore.repository.security.DisallowedException;
 import org.entrystore.repository.test.TestSuite;
 import org.entrystore.repository.util.NS;
 import org.entrystore.repository.util.URISplit;
-import org.openrdf.model.Graph;
-import org.openrdf.model.Literal;
-import org.openrdf.model.Resource;
-import org.openrdf.model.Statement;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.model.impl.ValueFactoryImpl;
-import org.openrdf.repository.RepositoryConnection;
-import org.openrdf.repository.RepositoryException;
-import org.openrdf.repository.RepositoryResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -64,6 +64,9 @@ import java.util.Map;
 import java.util.Set;
 import java.util.StringTokenizer;
 
+import static org.eclipse.rdf4j.model.util.Values.iri;
+import static org.eclipse.rdf4j.model.util.Values.literal;
+
 
 public class ContextImpl extends ResourceImpl implements Context {
 
@@ -75,8 +78,8 @@ public class ContextImpl extends ResourceImpl implements Context {
 	protected ArrayList<URI> systemEntries = new ArrayList<URI>();
 	private static Logger log = LoggerFactory.getLogger(ContextImpl.class);
 
-	public static final org.openrdf.model.URI DCModified;
-	public static final org.openrdf.model.URI DCTermsModified;
+	public static final IRI DCModified;
+	public static final IRI DCTermsModified;
 	
 	private Object quotaMutex = new Object();
 	protected long quotaFillLevel = Quota.VALUE_UNCACHED;
@@ -85,9 +88,8 @@ public class ContextImpl extends ResourceImpl implements Context {
 	private volatile boolean deleted = false;
 
 	static {
-		ValueFactory vf = ValueFactoryImpl.getInstance();
-		DCModified = vf.createURI(NS.dc, "modified");
-		DCTermsModified = vf.createURI(NS.dcterms, "modified");
+		DCModified = iri(NS.dc, "modified");
+		DCTermsModified = iri(NS.dcterms, "modified");
 	}
 
 	protected ContextImpl(EntryImpl entry, String uri, SoftCache cache) {
@@ -96,7 +98,7 @@ public class ContextImpl extends ResourceImpl implements Context {
 		this.id = uri.substring(uri.lastIndexOf('/')+1);
 	}
 
-	public ContextImpl(EntryImpl entry, org.openrdf.model.URI contextUri, SoftCache cache)  {
+	public ContextImpl(EntryImpl entry, IRI contextUri, SoftCache cache)  {
 		super(entry, contextUri);
 		this.cache = cache;
 		this.id = resourceURI.toString().substring(resourceURI.toString().lastIndexOf('/') + 1);
@@ -130,13 +132,13 @@ public class ContextImpl extends ResourceImpl implements Context {
 					while (resources.hasNext()) {
 						Statement statement = resources.next();
 						Resource mmd = statement.getContext();
-						if (mmd instanceof org.openrdf.model.URI) {
+						if (mmd instanceof IRI) {
 							if (!mmd.stringValue().startsWith(entry.getRepositoryManager().getRepositoryURL().toString())) {
 								log.warn("This Entry URI does not belong to this repository: " + mmd.stringValue());
 								continue;
 							}
 							
-							StringTokenizer stok = Util.extractParameters(entry.repositoryManager, (org.openrdf.model.URI) mmd);
+							StringTokenizer stok = Util.extractParameters(entry.repositoryManager, (IRI) mmd);
 							if (stok.countTokens() == 3 && stok.nextToken().equals(this.id)) { //Belongs to this context.
 								try {
 									stok.nextToken(); //Ignoring the M
@@ -156,8 +158,8 @@ public class ContextImpl extends ResourceImpl implements Context {
 					while (externalMD.hasNext()) {
 						Statement statement = externalMD.next();
 						Resource mmd = statement.getContext();
-						if (mmd instanceof org.openrdf.model.URI) {
-							StringTokenizer stok = Util.extractParameters(entry.repositoryManager, (org.openrdf.model.URI) mmd);
+						if (mmd instanceof IRI) {
+							StringTokenizer stok = Util.extractParameters(entry.repositoryManager, (IRI) mmd);
 							if (stok.countTokens() == 3 && stok.nextToken().equals(this.id)) { //Belongs to this context.
 								stmntsToAdd.add(vf.createStatement((Resource) statement.getObject(), RepositoryProperties.mdHasEntry, statement.getContext(), this.resourceURI));
 							}
@@ -264,7 +266,7 @@ public class ContextImpl extends ResourceImpl implements Context {
 					while (statements.hasNext()) {
 						Statement statement = statements.next();
 						try {
-							org.openrdf.model.URI predicate = statement.getPredicate();
+							IRI predicate = statement.getPredicate();
 							if (predicate.equals(RepositoryProperties.mdHasEntry)) {
 								URI mdURI = URI.create(statement.getSubject().toString());
 								URI entryURI = URI.create(statement.getObject().toString());
@@ -291,7 +293,7 @@ public class ContextImpl extends ResourceImpl implements Context {
 		}
 	}
 
-	private void addToIndex(org.openrdf.model.URI entryURI, org.openrdf.model.URI resURI, org.openrdf.model.URI extMdURI, RepositoryConnection rc) throws RepositoryException {		
+	private void addToIndex(IRI entryURI, IRI resURI, IRI extMdURI, RepositoryConnection rc) throws RepositoryException {
 		rc.add(resURI, RepositoryProperties.resHasEntry, entryURI, this.resourceURI);							
 		URI euri = URI.create(entryURI.toString());
 
@@ -309,9 +311,9 @@ public class ContextImpl extends ResourceImpl implements Context {
 	}
 
 	protected void removeFromIndex(EntryImpl entry, RepositoryConnection rc) throws RepositoryException {
-		org.openrdf.model.URI entryURI = entry.getSesameEntryURI();
-		org.openrdf.model.URI resURI = entry.getSesameResourceURI();
-		org.openrdf.model.URI mdURI = entry.getSesameExternalMetadataURI();
+		IRI entryURI = entry.getSesameEntryURI();
+		IRI resURI = entry.getSesameResourceURI();
+		IRI mdURI = entry.getSesameExternalMetadataURI();
 
 		rc.remove(resURI, RepositoryProperties.resHasEntry, entryURI, this.resourceURI);
 
@@ -340,7 +342,7 @@ public class ContextImpl extends ResourceImpl implements Context {
 			}
 			URI deletedBy = entry.getRepositoryManager().getPrincipalManager().getAuthenticatedUserURI();
 			if (deletedBy != null) {
-				Statement delByStmnt = vf.createStatement(entryURI, RepositoryProperties.DeletedBy, vf.createURI(deletedBy.toString()), this.resourceURI);
+				Statement delByStmnt = vf.createStatement(entryURI, RepositoryProperties.DeletedBy, vf.createIRI(deletedBy.toString()), this.resourceURI);
 				rc.add(delByStmnt, this.resourceURI);
 			}
 		}
@@ -455,23 +457,22 @@ public class ContextImpl extends ResourceImpl implements Context {
 					do {
 						counter++;
 						identity = Long.toString(counter);
-						org.openrdf.model.URI entryUri = vf.createURI(base 
-								+ this.id + "/" + RepositoryProperties.ENTRY_PATH + "/" + Long.toString(counter));
+						IRI entryUri = vf.createIRI(base + this.id + "/" + RepositoryProperties.ENTRY_PATH + "/" + Long.toString(counter));
 						infoRecord = rc.getStatements(null, null, null, false, entryUri).asList();
 					} while (!infoRecord.isEmpty()); //keep counting if candidate is taken
 				}
 
 				//resURI - resourceURI
-				org.openrdf.model.URI resURI = null;
+				IRI resURI = null;
 				if (resourceURI != null) {
 					String resourceURIStr = resourceURI.toString().replace("_newId", identity);
-					resURI = vf.createURI(resourceURIStr);
+					resURI = vf.createIRI(resourceURIStr);
 				} else {
 					if (bType == GraphType.Context ||
 							bType == GraphType.SystemContext) {
-						resURI = vf.createURI(URISplit.fabricateContextURI(base, identity).toString());					
+						resURI = vf.createIRI(URISplit.fabricateContextURI(base, identity).toString());
 					} else {
-						resURI = vf.createURI(URISplit.fabricateURI(base, this.id, RepositoryProperties.getResourcePath(bType), identity).toString());
+						resURI = vf.createIRI(URISplit.fabricateURI(base, this.id, RepositoryProperties.getResourcePath(bType), identity).toString());
 					}
 				}
 
@@ -482,7 +483,7 @@ public class ContextImpl extends ResourceImpl implements Context {
 
 					//Initialize a new information object.
 					if (lType == EntryType.Reference || lType == EntryType.LinkReference) {
-						newEntry.create(resURI, vf.createURI(metadataURI.toString()), bType, lType, rType, rc);
+						newEntry.create(resURI, vf.createIRI(metadataURI.toString()), bType, lType, rType, rc);
 					} else {
 						newEntry.create(resURI, null, bType, lType, rType, rc);					
 					}
@@ -1181,12 +1182,11 @@ public class ContextImpl extends ResourceImpl implements Context {
 
 	public void setMetadata(Entry entry, String title, String desc) {
 		try {
-			Graph graph = entry.getLocalMetadata().getGraph();
-			ValueFactory vf = graph.getValueFactory(); 
-			org.openrdf.model.URI root = vf.createURI(entry.getResourceURI().toString());
-			graph.add(root, TestSuite.dc_title, vf.createLiteral(title, "en"));
+			Model graph = entry.getLocalMetadata().getGraph();
+			IRI root = iri(entry.getResourceURI().toString());
+			graph.add(root, TestSuite.dc_title, literal(title, "en"));
 			if (desc != null) {
-				graph.add(root, TestSuite.dc_description, vf.createLiteral(desc, "en"));
+				graph.add(root, TestSuite.dc_description, literal(desc, "en"));
 			}
 			entry.getLocalMetadata().setGraph(graph);
 		} catch (Exception e) {
