@@ -16,43 +16,39 @@
 
 package org.entrystore.impl;
 
-import java.util.Iterator;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.Statement;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.model.vocabulary.RDF;
+import org.eclipse.rdf4j.model.vocabulary.XMLSchema;
 
-import org.openrdf.model.Graph;
-import org.openrdf.model.Statement;
-import org.openrdf.model.URI;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.model.impl.StatementImpl;
-import org.openrdf.model.vocabulary.RDF;
-import org.openrdf.model.vocabulary.XMLSchema;
+import java.util.Iterator;
 
 
 public class StringResource extends RDFResource{
 
-	public StringResource(EntryImpl entry, URI resourceURI) {
+	public StringResource(EntryImpl entry, IRI resourceURI) {
 		super(entry, resourceURI);
 	}
 
-	public StringResource(EntryImpl entry, URI resourceURI, String str) {
+	public StringResource(EntryImpl entry, IRI resourceURI, String str) {
 		super(entry, resourceURI);
 		setString(str); 
 	}
 
 	public void setString(String text) {
-		Graph g = getGraph();
+		Model g = getGraph();
 		ValueFactory vf = this.entry.repository.getValueFactory(); 
 		g.clear();
 		if (text != null && !text.equals("")) {
-			g.add(new StatementImpl(
-					this.entry.getSesameResourceURI(), 
-					RDF.VALUE,
-					vf.createLiteral(text, XMLSchema.STRING)));
+			g.add(vf.createStatement(this.entry.getSesameResourceURI(),	RDF.VALUE, vf.createLiteral(text, XMLSchema.STRING)));
 		}
 		this.setGraph(g); 
 	}
 
 	public String getString() {
-		Iterator<Statement> stringElements = this.getGraph().match(this.entry.getSesameResourceURI(), RDF.VALUE, null);
+		Iterator<Statement> stringElements = this.getGraph().filter(this.entry.getSesameResourceURI(), RDF.VALUE, null).iterator();
 		while(stringElements.hasNext()) {
 			return stringElements.next().getObject().stringValue(); 
 		}
