@@ -17,16 +17,17 @@
 
 package org.entrystore.impl;
 
-import java.util.StringTokenizer;
-
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.ValueFactory;
 import org.entrystore.repository.RepositoryManager;
-import org.openrdf.model.URI;
-import org.openrdf.model.ValueFactory;
+
+import java.net.URI;
+import java.util.StringTokenizer;
 
 
 public class Util {
 		
-	public static StringTokenizer extractParameters(RepositoryManager rman, URI uri) {
+	public static StringTokenizer extractParameters(RepositoryManager rman, IRI uri) {
 		return extractParameters(rman, uri.toString());
 	}
 
@@ -35,26 +36,26 @@ public class Util {
 		return new StringTokenizer(withoutBase, "/");
 	}
 	
-	public static URI getMetadataURI(EntryImpl item, String contextID, String uniqueID) {
+	public static IRI getMetadataURI(EntryImpl item, String contextID, String uniqueID) {
 		ValueFactory vf = item.getRepository().getValueFactory();
 		String base = item.getRepositoryManager().getRepositoryURL().toString();
-		return vf.createURI(base + contextID + "/" + RepositoryProperties.MD_PATH + "/" + uniqueID);
+		return vf.createIRI(base + contextID + "/" + RepositoryProperties.MD_PATH + "/" + uniqueID);
 	}
 
-	public static java.net.URI getMMdURIFromMdURI(URI md) {
-		return java.net.URI.create(md.toString().replaceFirst("/"+RepositoryProperties.MD_PATH+"/", "/"+RepositoryProperties.ENTRY_PATH+"/"));
+	public static URI getMMdURIFromMdURI(IRI md) {
+		return URI.create(md.toString().replaceFirst("/"+RepositoryProperties.MD_PATH+"/", "/"+RepositoryProperties.ENTRY_PATH+"/"));
 	}
 
-	public static String getContextIdFromURI(RepositoryManager rman, java.net.URI sourceURI) {
+	public static String getContextIdFromURI(RepositoryManager rman, URI sourceURI) {
 		return extractParameters(rman, sourceURI.toString()).nextToken();
 	}
 	
-	public static java.net.URI getContextURIFromURI(RepositoryManager rman, java.net.URI sourceURI) {
-		return java.net.URI.create(rman.getRepositoryURL().toString()+ "/" + RepositoryProperties.SYSTEM_CONTEXTS_ID + "/" 
+	public static URI getContextURIFromURI(RepositoryManager rman, URI sourceURI) {
+		return URI.create(rman.getRepositoryURL().toString()+ "/" + RepositoryProperties.SYSTEM_CONTEXTS_ID + "/" 
 				+ RepositoryProperties.DATA_PATH + "/" + getContextIdFromURI(rman, sourceURI));
 	}
-	public static java.net.URI getContextMMdURIFromURI(RepositoryManager rman, java.net.URI sourceURI) {
-		return java.net.URI.create(rman.getRepositoryURL().toString()+ "/" + RepositoryProperties.SYSTEM_CONTEXTS_ID + "/" 
+	public static URI getContextMMdURIFromURI(RepositoryManager rman, URI sourceURI) {
+		return URI.create(rman.getRepositoryURL().toString()+ "/" + RepositoryProperties.SYSTEM_CONTEXTS_ID + "/" 
 				+ RepositoryProperties.ENTRY_PATH + "/" + getContextIdFromURI(rman, sourceURI));
 	}
 }
