@@ -16,20 +16,24 @@
 
 package org.entrystore.impl.converters;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Literal;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.impl.LinkedHashModel;
+import org.entrystore.Converter;
+import org.w3c.dom.NamedNodeMap;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+
+import java.net.URI;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.entrystore.Converter;
-import org.openrdf.model.Graph;
-import org.openrdf.model.Literal;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.model.impl.GraphImpl;
-import org.w3c.dom.NamedNodeMap;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
+import static org.eclipse.rdf4j.model.util.Values.iri;
+import static org.eclipse.rdf4j.model.util.Values.literal;
 
 
 public class OAI_DC2RDFGraphConverter implements Converter {
@@ -58,7 +62,7 @@ public class OAI_DC2RDFGraphConverter implements Converter {
 	 * 
 	 * @return the new metadata graph.
 	 */
-	public Object convert(Object from, java.net.URI resourceURI, java.net.URI metadataURI) {
+	public Object convert(Object from, URI resourceURI, URI metadataURI) {
 		NodeList metadataList = null;
 
 		if (from instanceof NodeList) {
@@ -70,9 +74,8 @@ public class OAI_DC2RDFGraphConverter implements Converter {
 			return null;
 		}
 
-		Graph graph = new GraphImpl();
-		ValueFactory vf = graph.getValueFactory();
-		org.openrdf.model.URI root = vf.createURI(resourceURI.toString());
+		Model graph = new LinkedHashModel();
+		IRI root = iri(resourceURI.toString());
 
 		for (int i = 0; i < metadataList.getLength(); i++) {
 			Node n = metadataList.item(i);
@@ -125,12 +128,12 @@ public class OAI_DC2RDFGraphConverter implements Converter {
 			
 			Literal lit;
 			if (lang != null) {
-				lit = vf.createLiteral(nodeContent, lang);
+				lit = literal(nodeContent, lang);
 			} else {
-				lit = vf.createLiteral(nodeContent);
+				lit = literal(nodeContent);
 			}
 			
-			graph.add(root, new org.openrdf.model.impl.URIImpl(predicate), lit);
+			graph.add(root, iri(predicate), lit);
 		}
 
 		return graph;
