@@ -16,6 +16,22 @@
 
 package org.entrystore.repository.transformation;
 
+import org.eclipse.rdf4j.model.BNode;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.Resource;
+import org.eclipse.rdf4j.model.Statement;
+import org.eclipse.rdf4j.model.Value;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.model.impl.LinkedHashModel;
+import org.eclipse.rdf4j.model.impl.SimpleValueFactory;
+import org.eclipse.rdf4j.model.vocabulary.RDF;
+import org.eclipse.rdf4j.repository.sail.SailRepository;
+import org.eclipse.rdf4j.rio.RDFParser;
+import org.eclipse.rdf4j.rio.helpers.BasicParserSettings;
+import org.eclipse.rdf4j.rio.helpers.StatementCollector;
+import org.eclipse.rdf4j.rio.n3.N3ParserFactory;
+import org.eclipse.rdf4j.sail.memory.MemoryStore;
 import org.entrystore.Context;
 import org.entrystore.Data;
 import org.entrystore.Entry;
@@ -24,36 +40,21 @@ import org.entrystore.QuotaException;
 import org.entrystore.ResourceType;
 import org.entrystore.impl.EntryImpl;
 import org.entrystore.repository.RepositoryException;
-import org.openrdf.model.BNode;
-import org.openrdf.model.Graph;
-import org.openrdf.model.Resource;
-import org.openrdf.model.Statement;
-import org.openrdf.model.URI;
-import org.openrdf.model.Value;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.model.impl.GraphImpl;
-import org.openrdf.model.impl.ValueFactoryImpl;
-import org.openrdf.model.vocabulary.RDF;
-import org.openrdf.repository.sail.SailRepository;
-import org.openrdf.rio.RDFHandlerException;
-import org.openrdf.rio.RDFParseException;
-import org.openrdf.rio.RDFParser;
-import org.openrdf.rio.helpers.StatementCollector;
-import org.openrdf.rio.n3.N3ParserFactory;
-import org.openrdf.sail.memory.MemoryStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URI;
 import java.nio.file.Files;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Vector;
+
+import static org.eclipse.rdf4j.model.util.Values.bnode;
 
 
 public class SCAM2Import {
@@ -64,25 +65,25 @@ public class SCAM2Import {
 	private Context context;
 	private File backup;
 	private SailRepository repository;
-	private Map<URI, Entry> uri2Entry = new HashMap<URI, Entry>();
+	private Map<IRI, Entry> uri2Entry = new HashMap<IRI, Entry>();
 	private File files;
 
-	public static final URI manifest;
-	public static final URI organizations;
-	public static final URI firstChild;
-	public static final URI KMRcollection;
-	public static final URI KMRURL;
-	public static final URI KMRFile;
-	public static final URI KMRDisplayName;
-	public static final URI KMRrecord;
-	public static final URI IMScontent;
-	public static final URI IMSItem;
-	public static final URI LOMTlocation;
-	public static final URI DCTitle;
-	public static final URI DCDescription;
-	public static final URI DCTTitle;
-	public static final URI DCTDescription;
-	public static final URI RDF_1;
+	public static final IRI manifest;
+	public static final IRI organizations;
+	public static final IRI firstChild;
+	public static final IRI KMRcollection;
+	public static final IRI KMRURL;
+	public static final IRI KMRFile;
+	public static final IRI KMRDisplayName;
+	public static final IRI KMRrecord;
+	public static final IRI IMScontent;
+	public static final IRI IMSItem;
+	public static final IRI LOMTlocation;
+	public static final IRI DCTitle;
+	public static final IRI DCDescription;
+	public static final IRI DCTTitle;
+	public static final IRI DCTDescription;
+	public static final IRI RDF_1;
 	
 	
 	public static final String NSbase = "http://www.imsproject.org/xsd/ims_cp_rootv1p1#";
@@ -94,23 +95,23 @@ public class SCAM2Import {
 	
 	
 	static {
-		ValueFactory vf = ValueFactoryImpl.getInstance();
-		manifest = vf.createURI(NSbase + "Manifest");
-		organizations = vf.createURI(NSbase + "organizations");
-		firstChild = vf.createURI(RDF.NAMESPACE+"_1");
-		KMRcollection = vf.createURI(KMRbase +"/datatypes#Collection");
-		KMRURL = vf.createURI(KMRbase +"#Url");
-		KMRFile = vf.createURI(KMRbase +"#File");
-		KMRDisplayName = vf.createURI(KMRbase +"#displayName");
-		KMRrecord = vf.createURI(KMRbase +"#record");
-		IMScontent = vf.createURI(IMSbase +"content");
-		IMSItem = vf.createURI(IMSbase +"Item");
-		LOMTlocation = vf.createURI(LOMTechnicalbase+"location");
-		DCTitle = vf.createURI(DCbase+"title");
-		DCDescription = vf.createURI(DCbase+"description");
-		DCTTitle = vf.createURI(DCTermsbase+"title");
-		DCTDescription = vf.createURI(DCTermsbase+"description");
-		RDF_1 = vf.createURI(RDF.NAMESPACE+"_1");
+		ValueFactory vf = SimpleValueFactory.getInstance();
+		manifest = vf.createIRI(NSbase + "Manifest");
+		organizations = vf.createIRI(NSbase + "organizations");
+		firstChild = vf.createIRI(RDF.NAMESPACE+"_1");
+		KMRcollection = vf.createIRI(KMRbase +"/datatypes#Collection");
+		KMRURL = vf.createIRI(KMRbase +"#Url");
+		KMRFile = vf.createIRI(KMRbase +"#File");
+		KMRDisplayName = vf.createIRI(KMRbase +"#displayName");
+		KMRrecord = vf.createIRI(KMRbase +"#record");
+		IMScontent = vf.createIRI(IMSbase +"content");
+		IMSItem = vf.createIRI(IMSbase +"Item");
+		LOMTlocation = vf.createIRI(LOMTechnicalbase+"location");
+		DCTitle = vf.createIRI(DCbase+"title");
+		DCDescription = vf.createIRI(DCbase+"description");
+		DCTTitle = vf.createIRI(DCTermsbase+"title");
+		DCTDescription = vf.createIRI(DCTermsbase+"description");
+		RDF_1 = vf.createIRI(RDF.NAMESPACE+"_1");
 	}
 
 	public SCAM2Import(Context importToContext, String pathToScam2BackupDir) {
@@ -122,7 +123,7 @@ public class SCAM2Import {
 	
 	public void doImport() {
 		RDFParser parser = new N3ParserFactory().getParser();
-		parser.setDatatypeHandling(RDFParser.DatatypeHandling.IGNORE);
+		parser.getParserConfig().set(BasicParserSettings.VERIFY_DATATYPE_VALUES, false);
 
 		StatementCollector collector = new StatementCollector();
 		parser.setRDFHandler(collector);
@@ -130,30 +131,22 @@ public class SCAM2Import {
 			InputStream fileOut = Files.newInputStream(new File(this.backup, "manifest.n3").toPath());
 			parser.parse(fileOut, "");
 			fileOut.close();
-			Graph graph = new GraphImpl(collector.getStatements());
-			Resource manifestRes = graph.match(null, RDF.TYPE, manifest).next().getSubject();
-			Resource organizationsRes = (Resource) graph.match(manifestRes, organizations, null).next().getObject();
-			URI first = getFirstChild(graph, organizationsRes);
+			Model graph = new LinkedHashModel(collector.getStatements());
+			Resource manifestRes = graph.filter(null, RDF.TYPE, manifest).iterator().next().getSubject();
+			Resource organizationsRes = (Resource) graph.filter(manifestRes, organizations, null).iterator().next().getObject();
+			IRI first = getFirstChild(graph, organizationsRes);
 			recurse(graph, null, first);
 			Entry top = context.get("_top");
 			Entry firstEntry = uri2Entry.get(first);
 			((org.entrystore.List) top.getResource()).addChild(firstEntry.getEntryURI());
 			recurseFix(graph, first, (org.entrystore.List) firstEntry.getResource());
-		} catch (FileNotFoundException e) {
-			e.printStackTrace();
-		} catch (RDFHandlerException e) {
-			e.printStackTrace();
-		} catch (RDFParseException e) {
-			e.printStackTrace();
-		} catch (IOException e) {
-			e.printStackTrace();
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
 	}
 
-	void recurseFix(Graph graph, URI folder, org.entrystore.List list) {
-		for (URI child : getChildren(graph, folder)) {
+	void recurseFix(Model graph, IRI folder, org.entrystore.List list) {
+		for (IRI child : getChildren(graph, folder)) {
 			Entry entryChild = uri2Entry.get(child);
 			if (entryChild != null) {
 				try {
@@ -164,7 +157,7 @@ public class SCAM2Import {
 					recurseFix(graph, child, (org.entrystore.List) entryChild.getResource());
 				}
 			} else {
-				URI refChild = getContent(graph, child);
+				IRI refChild = getContent(graph, child);
 				Entry refEntryChild = uri2Entry.get(refChild);
 				if (refEntryChild != null) {
 					try {
@@ -176,13 +169,13 @@ public class SCAM2Import {
 		}
 	}
 
-	void recurse(Graph graph, URI parent, URI folder) {
+	void recurse(Model graph, IRI parent, IRI folder) {
 		handleFolder(graph, parent, folder);
-		for (URI child : getChildren(graph, folder)) {
+		for (IRI child : getChildren(graph, folder)) {
 			if (isFolder(graph, child)) {
 				recurse(graph, folder, child);
 			} else {
-				URI refChild = getContent(graph, child);
+				IRI refChild = getContent(graph, child);
 				if (!isItem(graph, refChild) &&
 						!uri2Entry.containsKey(refChild)) {
 					handleLeaf(graph, folder, refChild);
@@ -191,23 +184,23 @@ public class SCAM2Import {
 		}
 	}
 	
-	void populateFolder(Graph graph, URI parent, URI folder) {
+	void populateFolder(Model graph, IRI parent, IRI folder) {
 	}
 	
-	void handleFolder(Graph graph, URI parent, URI folder) {
-		Graph closure = getAnonymousClosure(graph, folder);
-		java.net.URI parentList = parent == null ? null : uri2Entry.get(parent).getResourceURI();
+	void handleFolder(Model graph, IRI parent, IRI folder) {
+		Model closure = getAnonymousClosure(graph, folder);
+		URI parentList = parent == null ? null : uri2Entry.get(parent).getResourceURI();
 		Entry folderEntry = context.createResource(null, GraphType.List, ResourceType.InformationResource, parentList);
 		handleItem((EntryImpl) folderEntry, closure, parent, folder);
 	}
 
-	void handleLeaf(Graph graph, URI parent, URI leaf) {
+	void handleLeaf(Model graph, IRI parent, IRI leaf) {
 		log.warn("Working with \""+leaf.stringValue()+"\"");
-		Graph closure = getAnonymousClosure(graph, leaf);
-		java.net.URI parentList = parent == null ? null : uri2Entry.get(parent).getResourceURI();
+		Model closure = getAnonymousClosure(graph, leaf);
+		URI parentList = parent == null ? null : uri2Entry.get(parent).getResourceURI();
 		Entry leafEntry = null;
 		if(isURL(closure, leaf)) {
-			leafEntry = context.createLink(null, java.net.URI.create(leaf.stringValue()), parentList);
+			leafEntry = context.createLink(null, URI.create(leaf.stringValue()), parentList);
 		} else if (isFile(closure, leaf)) {
 			leafEntry = context.createResource(null, GraphType.None, ResourceType.InformationResource, parentList);
 			setFile((Data) leafEntry.getResource(), closure, leaf);
@@ -219,13 +212,13 @@ public class SCAM2Import {
 		}
 	}
 
-	void handleItem(EntryImpl entry, Graph closure, URI parent, URI item) {
+	void handleItem(EntryImpl entry, Model closure, IRI parent, IRI item) {
 		uri2Entry.put(item, entry);
-		Graph metadata = new GraphImpl();
+		Model metadata = new LinkedHashModel();
 		for (Statement statement : closure) {
 			Resource s = entry.getSesameResourceURI();
 			Resource subject = statement.getSubject();
-			URI predicate = statement.getPredicate();
+			IRI predicate = statement.getPredicate();
 			String object = statement.getObject().stringValue();
 			if (object.startsWith("urn:x-")) {
 				continue;
@@ -243,13 +236,12 @@ public class SCAM2Import {
 				//Taken care of separately.
 			} else if (predicate.equals(KMRDisplayName)) {
 				entry.setFilename(statement.getObject().stringValue());
-			} else if (subject instanceof URI){ //Works since no other URIs are in subject position in a anonymous closure
+			} else if (subject instanceof IRI){ //Works since no other URIs are in subject position in a anonymous closure
 				if (!predicate.stringValue().startsWith(RDF.NAMESPACE+"_")) { //Ignore all rdf-collection relations.
 					if (predicate.equals(DCTitle)) {
 						metadata.add(s, DCTTitle, statement.getObject());						
 					} else if (predicate.equals(DCDescription)) {
-						ValueFactory vf = ValueFactoryImpl.getInstance();
-						BNode bnode = vf.createBNode();
+						BNode bnode = bnode();
 						metadata.add(s, DCTDescription, bnode);
 						metadata.add(bnode, RDF.VALUE, statement.getObject());						
 					} else {
@@ -264,8 +256,8 @@ public class SCAM2Import {
 		entry.getLocalMetadata().setGraph(metadata);
 	}
 
-	void setFile(Data data, Graph closure, URI uri) {
-		URI loc = getLocation(closure, uri);
+	void setFile(Data data, Model closure, IRI uri) {
+		IRI loc = getLocation(closure, uri);
 		if (loc != null) {
 			String locStr = loc.stringValue();
 			locStr = locStr.substring(locStr.indexOf("&uri=")+5);
@@ -284,53 +276,51 @@ public class SCAM2Import {
 		}
 	}
 	
-	Graph getAnonymousClosure(Graph graph, URI subject) {
-		Graph collect = new GraphImpl();
+	Model getAnonymousClosure(Model graph, IRI subject) {
+		Model collect = new LinkedHashModel();
 		collectAnonymousClosure(graph, subject, collect);
 		return collect;
 	}
 
 	
-	void collectAnonymousClosure(Graph graph, Resource subject, Graph collect) {
-		Iterator<Statement> it = graph.match(subject, null, null);
-		while(it.hasNext()) {
-			Statement st = it.next();
+	void collectAnonymousClosure(Model graph, Resource subject, Model collect) {
+		for (Statement st : graph.filter(subject, null, null)) {
 			Value object = st.getObject();
 			if (st.getPredicate().equals(KMRrecord)) {
 				continue;
 			}
 			collect.add(st);
 			//If blank
-			if (object instanceof Resource 
-					&& !(object instanceof URI)) {
+			if (object instanceof Resource
+					&& !(object instanceof IRI)) {
 				collectAnonymousClosure(graph, (Resource) object, collect);
 			}
 		}
 	}
 	
-	boolean isFolder(Graph graph, Resource subject) {
-		return graph.match(subject, RDF_1, null).hasNext();
+	boolean isFolder(Model graph, Resource subject) {
+		return graph.filter(subject, RDF_1, null).iterator().hasNext();
 		//return graph.match(subject, RDF.TYPE, KMRcollection).hasNext();
 	}
 
-	boolean isItem(Graph graph, Resource subject) {
-		return graph.match(subject, RDF.TYPE, IMSItem).hasNext();
+	boolean isItem(Model graph, Resource subject) {
+		return graph.filter(subject, RDF.TYPE, IMSItem).iterator().hasNext();
 	}
 
-	boolean isURL(Graph graph, Resource subject) {
-		return graph.match(subject, RDF.TYPE, KMRURL).hasNext();
+	boolean isURL(Model graph, Resource subject) {
+		return graph.filter(subject, RDF.TYPE, KMRURL).iterator().hasNext();
 	}
 
-	boolean isFile(Graph graph, Resource subject) {
-		return graph.match(subject, RDF.TYPE, KMRFile).hasNext();
+	boolean isFile(Model graph, Resource subject) {
+		return graph.filter(subject, RDF.TYPE, KMRFile).iterator().hasNext();
 	}
 
-	URI getContent(Graph graph, URI item) {
-		Iterator<Statement> it = graph.match(item, IMScontent, null);
+	IRI getContent(Model graph, IRI item) {
+		Iterator<Statement> it = graph.filter(item, IMScontent, null).iterator();
 		if (it.hasNext()) {
 			Value obj = it.next().getObject();
-			if (obj instanceof URI) {
-				return (URI) obj;
+			if (obj instanceof IRI) {
+				return (IRI) obj;
 			} else {
 				return null;
 			}
@@ -338,12 +328,12 @@ public class SCAM2Import {
 		return null;
 	}
 
-	URI getLocation(Graph graph, URI item) {
-		Iterator<Statement> it = graph.match(item, LOMTlocation, null);
+	IRI getLocation(Model graph, IRI item) {
+		Iterator<Statement> it = graph.filter(item, LOMTlocation, null).iterator();
 		if (it.hasNext()) {
 			Value obj = it.next().getObject();
-			if (obj instanceof URI) {
-				return (URI) obj;
+			if (obj instanceof IRI) {
+				return (IRI) obj;
 			} else {
 				return null;
 			}
@@ -351,12 +341,12 @@ public class SCAM2Import {
 		return null;	
 	}
 
-	URI getFirstChild(Graph graph, Resource subject) {
-		Iterator<Statement> it = graph.match(subject, firstChild, null);
+	IRI getFirstChild(Model graph, Resource subject) {
+		Iterator<Statement> it = graph.filter(subject, firstChild, null).iterator();
 		if (it.hasNext()) {
 			Value obj = it.next().getObject();
-			if (obj instanceof URI) {
-				return (URI) obj;
+			if (obj instanceof IRI) {
+				return (IRI) obj;
 			} else {
 				return null;
 			}
@@ -364,23 +354,20 @@ public class SCAM2Import {
 		return null;
 	}
 	
-	List<URI> getChildren(Graph graph, Resource subject) {
-		Vector<URI> children = new Vector<URI>();
-		Iterator<Statement> sts = graph.match(subject, null, null);
-		while (sts.hasNext()) {
-			Statement st = sts.next();
+	List<IRI> getChildren(Model graph, Resource subject) {
+		Vector<IRI> children = new Vector<IRI>();
+		for (Statement st : graph.filter(subject, null, null)) {
 			try {
 				String value = st.getPredicate().toString().substring(RDF.NAMESPACE.length());
-				int index = Integer.parseInt(value.substring(value.lastIndexOf("_")+1));					
+				int index = Integer.parseInt(value.substring(value.lastIndexOf("_") + 1));
 				if (index > children.size()) {
-					children.setSize(index); 
+					children.setSize(index);
 				}
 
-				if (st.getObject() instanceof URI) {
-					children.set(index-1, (URI) st.getObject());
+				if (st.getObject() instanceof IRI) {
+					children.set(index - 1, (IRI) st.getObject());
 				}
-			} catch (IndexOutOfBoundsException iobe) {
-			} catch (NumberFormatException nfe) {
+			} catch (IndexOutOfBoundsException | NumberFormatException iobe) {
 			}
 		}
 		children.trimToSize();
