@@ -16,10 +16,10 @@
 
 package org.entrystore.repository.util;
 
+import org.eclipse.rdf4j.query.MalformedQueryException;
+import org.eclipse.rdf4j.repository.RepositoryException;
 import org.entrystore.ContextManager;
 import org.entrystore.Entry;
-import org.openrdf.query.MalformedQueryException;
-import org.openrdf.repository.RepositoryException;
 
 
 public class CommonQueries {
