@@ -16,6 +16,11 @@
 
 package org.entrystore.impl;
 
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.query.MalformedQueryException;
+import org.eclipse.rdf4j.repository.RepositoryException;
 import org.entrystore.Context;
 import org.entrystore.Entry;
 import org.entrystore.EntryType;
@@ -24,10 +29,6 @@ import org.entrystore.repository.util.CommonQueries;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
-import org.openrdf.model.Graph;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.query.MalformedQueryException;
-import org.openrdf.repository.RepositoryException;
 
 import java.net.URI;
 import java.util.ArrayList;
@@ -56,24 +57,23 @@ public class ContextManagerImplTest extends AbstractCoreTest {
 		Entry linkEntry = context.createLink(null, URI.create("http://slashdot.org/"), null);
 		Entry refEntry = context.createReference(null, URI.create("http://reddit.com/"), URI.create("http://example.com/md1"), null);
 
-		Graph graph = listEntry.getLocalMetadata().getGraph();
-		ValueFactory vf = graph.getValueFactory();
-		org.openrdf.model.URI root = vf.createURI(listEntry.getResource().getURI().toString());
+		Model graph = listEntry.getLocalMetadata().getGraph();
+		ValueFactory vf = rm.getValueFactory();
+		IRI root = vf.createIRI(listEntry.getResource().getURI().toString());
 
 
-		graph.add(root, new org.openrdf.model.impl.URIImpl("http://purl.org/dc/terms/title"), vf.createLiteral("Folder 1", "en"));
-		graph.add(root, new org.openrdf.model.impl.URIImpl("dc:description"), vf.createLiteral("A top level folder", "en"));
-		graph.add(root, new org.openrdf.model.impl.URIImpl("dc:subject"), vf.createLiteral("mainFolder"));
+		graph.add(root, vf.createIRI("http://purl.org/dc/terms/title"), vf.createLiteral("Folder 1", "en"));
+		graph.add(root, vf.createIRI("dc:description"), vf.createLiteral("A top level folder", "en"));
+		graph.add(root, vf.createIRI("dc:subject"), vf.createLiteral("mainFolder"));
 		listEntry.getLocalMetadata().setGraph(graph);
 
 
 		graph = linkEntry.getLocalMetadata().getGraph();
-		vf = graph.getValueFactory();
-		root = vf.createURI(linkEntry.getResourceURI().toString());
+		root = vf.createIRI(linkEntry.getResourceURI().toString());
 
-		graph.add(root, new org.openrdf.model.impl.URIImpl("http://purl.org/dc/terms/title"), vf.createLiteral("Dagens Nyheter"));
-		graph.add(root, new org.openrdf.model.impl.URIImpl("dc:description"), vf.createLiteral("A widely spread morning newspaper in sweden."));
-		graph.add(root, new org.openrdf.model.impl.URIImpl("dc:format"), vf.createLiteral("text/html"));
+		graph.add(root, vf.createIRI("http://purl.org/dc/terms/title"), vf.createLiteral("Dagens Nyheter"));
+		graph.add(root, vf.createIRI("dc:description"), vf.createLiteral("A widely spread morning newspaper in sweden."));
+		graph.add(root, vf.createIRI("dc:format"), vf.createLiteral("text/html"));
 		linkEntry.getLocalMetadata().setGraph(graph);
 
 		String from = "2008-06-01";
@@ -95,7 +95,7 @@ public class ContextManagerImplTest extends AbstractCoreTest {
 		Entry en = c.get("2");
 
 		if (!EntryType.Reference.equals(en.getEntryType())) {
-			Graph g = en.getLocalMetadata().getGraph();
+			Model g = en.getLocalMetadata().getGraph();
 			// g = en.getGraph();
 		}
 
@@ -138,7 +138,8 @@ public class ContextManagerImplTest extends AbstractCoreTest {
 
 		//Remove success?
 		try {
-			cm.remove(URI.create("http://exampls.com/nonexistingMMDURI")); //Should go wrong.
+			// should go wrong and will produce an ugly stack trace in the testing output, even though we catch it below
+			cm.remove(URI.create("http://exampls.com/nonexistingMMDURI"));
 			assertTrue(false);
 		} catch (Exception e) {
 		}
