@@ -16,25 +16,14 @@
 
 package org.entrystore.rest.auth;
 
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.model.impl.SimpleValueFactory;
 import org.entrystore.Entry;
-import org.entrystore.config.Config;
-import org.entrystore.repository.config.ConfigurationManager;
-import org.entrystore.repository.config.Settings;
 import org.entrystore.repository.util.NS;
-import org.entrystore.rest.util.Email;
-import org.openrdf.model.Graph;
-import org.openrdf.model.ValueFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.File;
-import java.io.IOException;
-import java.net.URI;
-import java.nio.ByteBuffer;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Calendar;
 
 /**
  * @author Hannes Ebner
@@ -44,13 +33,13 @@ public class Signup {
 	private static Logger log = LoggerFactory.getLogger(Signup.class);
 
 	public static void setFoafMetadata(Entry entry, org.restlet.security.User userInfo) {
-		Graph graph = entry.getLocalMetadata().getGraph();
-		ValueFactory vf = graph.getValueFactory();
-		org.openrdf.model.URI resourceURI = vf.createURI(entry.getResourceURI().toString());
+		Model graph = entry.getLocalMetadata().getGraph();
+		ValueFactory vf = SimpleValueFactory.getInstance();
+		IRI resourceURI = vf.createIRI(entry.getResourceURI().toString());
 		String fullname = null;
 		if (userInfo.getFirstName() != null) {
 			fullname = userInfo.getFirstName();
-			graph.add(vf.createStatement(resourceURI, vf.createURI(NS.foaf, "givenName"), vf.createLiteral(userInfo.getFirstName())));
+			graph.add(vf.createStatement(resourceURI, vf.createIRI(NS.foaf, "givenName"), vf.createLiteral(userInfo.getFirstName())));
 		}
 		if (userInfo.getLastName() != null) {
 			if (fullname != null) {
@@ -58,13 +47,13 @@ public class Signup {
 			} else {
 				fullname = userInfo.getLastName();
 			}
-			graph.add(vf.createStatement(resourceURI, vf.createURI(NS.foaf, "familyName"), vf.createLiteral(userInfo.getLastName())));
+			graph.add(vf.createStatement(resourceURI, vf.createIRI(NS.foaf, "familyName"), vf.createLiteral(userInfo.getLastName())));
 		}
 		if (fullname != null) {
-			graph.add(vf.createStatement(resourceURI, vf.createURI(NS.foaf, "name"), vf.createLiteral(fullname)));
+			graph.add(vf.createStatement(resourceURI, vf.createIRI(NS.foaf, "name"), vf.createLiteral(fullname)));
 		}
 		if (userInfo.getEmail() != null) {
-			graph.add(vf.createStatement(resourceURI, vf.createURI(NS.foaf, "mbox"), vf.createURI("mailto:", userInfo.getEmail())));
+			graph.add(vf.createStatement(resourceURI, vf.createIRI(NS.foaf, "mbox"), vf.createIRI("mailto:", userInfo.getEmail())));
 		}
 
 		entry.getLocalMetadata().setGraph(graph);
