@@ -19,11 +19,11 @@ package org.entrystore.rest.resources;
 import org.apache.commons.fileupload.FileItem;
 import org.apache.commons.fileupload.FileUploadException;
 import org.apache.commons.fileupload.disk.DiskFileItemFactory;
+import org.eclipse.rdf4j.repository.RepositoryException;
 import org.entrystore.AuthorizationException;
 import org.entrystore.PrincipalManager.AccessProperty;
 import org.entrystore.repository.transformation.SCAM2Import;
 import org.entrystore.repository.util.FileOperations;
-import org.openrdf.repository.RepositoryException;
 import org.restlet.Request;
 import org.restlet.data.MediaType;
 import org.restlet.data.Status;
