@@ -16,6 +16,14 @@
 
 package org.entrystore.rest.resources;
 
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Literal;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.Resource;
+import org.eclipse.rdf4j.model.Statement;
+import org.eclipse.rdf4j.model.Value;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.model.vocabulary.RDF;
 import org.entrystore.AuthorizationException;
 import org.entrystore.Context;
 import org.entrystore.ContextManager;
@@ -31,15 +39,6 @@ import org.entrystore.rest.util.JSONErrorMessages;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.openrdf.model.Graph;
-import org.openrdf.model.Literal;
-import org.openrdf.model.Resource;
-import org.openrdf.model.Statement;
-import org.openrdf.model.Value;
-import org.openrdf.model.impl.URIImpl;
-import org.openrdf.model.util.GraphUtil;
-import org.openrdf.model.util.GraphUtilException;
-import org.openrdf.model.vocabulary.RDF;
 import org.restlet.data.MediaType;
 import org.restlet.data.Status;
 import org.restlet.ext.json.JsonRepresentation;
@@ -56,10 +55,11 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
+import static org.eclipse.rdf4j.model.util.Values.iri;
 
 
 public class StatisticsResource extends BaseResource {
@@ -155,6 +155,7 @@ public class StatisticsResource extends BaseResource {
 	
 	public JSONObject getPropertyStatistics(Context context) {
 		JSONObject result = new JSONObject();
+		ValueFactory vf = getRM().getValueFactory();
 		
 		URI currentUserURI = getPM().getAuthenticatedUserURI();
 		getPM().setAuthenticatedUserURI(getPM().getAdminUser().getURI());
@@ -188,24 +189,23 @@ public class StatisticsResource extends BaseResource {
 				Set<String> entryProperties = new HashSet<String>();
 				String entryID = entry.getId();
 				URI resourceURI = entry.getResourceURI();
-				Graph metadata = entry.getMetadataGraph();
-				Iterator<Statement> firstLevel = metadata.match(new URIImpl(resourceURI.toString()), null, null);
-				while (firstLevel.hasNext()) {
-					org.openrdf.model.URI predicate = firstLevel.next().getPredicate();
+				Model metadata = entry.getMetadataGraph();
+				for (Statement statement : metadata.filter(vf.createIRI(resourceURI.toString()), null, null)) {
+					IRI predicate = statement.getPredicate();
 					String predStr = predicate.toString();
 					if (labelMap != null && labelMap.containsKey(predStr)) {
 						predStr = labelMap.get(predStr);
 					}
-					
+
 					entryProperties.add(predStr);
-					
+
 					// propertyUsage
 					int count = 0;
 					if (propertyUsage.containsKey(predStr)) {
 						count = propertyUsage.get(predStr);
 					}
 					propertyUsage.put(predStr, ++count);
-					
+
 					// usedIn
 					if (propertyUsedIn.containsKey(predStr)) {
 						Set<String> usedIn = propertyUsedIn.get(predStr);
@@ -263,6 +263,7 @@ public class StatisticsResource extends BaseResource {
 	
 	public JSONObject getOntologyStatistics(Context context) {
 		JSONObject result = new JSONObject();
+		ValueFactory vf = getRM().getValueFactory();
 		
 		URI currentUserURI = getPM().getAuthenticatedUserURI();
 		getPM().setAuthenticatedUserURI(getPM().getAdminUser().getURI());
@@ -284,61 +285,59 @@ public class StatisticsResource extends BaseResource {
 					continue;
 				}
 				
-				Set<org.openrdf.model.URI> allowedPredicates = new HashSet<org.openrdf.model.URI>();
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#Supports"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesAlternativeViewOn"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesExamplesOn"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#Methodology"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#Summarizes"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesDataOn"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesBackgroundOn"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#Details"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#IsAbout"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesTheoreticalInformationOn"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesNewInformationOn"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#CommentsOn"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#Refutes"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#Explains"));
-				allowedPredicates.add(new URIImpl("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesExamplesOf"));
+				Set<IRI> allowedPredicates = new HashSet<>();
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#Supports"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesAlternativeViewOn"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesExamplesOn"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#Methodology"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#Summarizes"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesDataOn"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesBackgroundOn"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#Details"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#IsAbout"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesTheoreticalInformationOn"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesNewInformationOn"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#CommentsOn"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#Refutes"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#Explains"));
+				allowedPredicates.add(vf.createIRI("http://www.cc.uah.es/ie/ont/OE-Predicates#ProvidesExamplesOf"));
 				
 				entryCount++;
 				Set<String> ontPredicates = new HashSet<String>();
 				Set<String> ontTerms = new HashSet<String>();
 				String entryID = entry.getId();
 				URI resourceURI = entry.getResourceURI();
-				Graph metadata = entry.getMetadataGraph();
-				Iterator<Statement> firstLevel = metadata.match(new URIImpl(resourceURI.toString()), null, null);
-				while (firstLevel.hasNext()) {
-					Statement stmnt = firstLevel.next();
-					org.openrdf.model.URI predicate = stmnt.getPredicate();
-					
+				Model metadata = entry.getMetadataGraph();
+				for (Statement stmnt : metadata.filter(vf.createIRI(resourceURI.toString()), null, null)) {
+					IRI predicate = stmnt.getPredicate();
+
 					if (!allowedPredicates.contains(predicate)) {
 						continue;
 					}
-					
+
 					entriesWithOntTerm.add(resourceURI);
-					
+
 					String predStr = predicate.stringValue().substring(predicate.stringValue().lastIndexOf("#") + 1);
 					Value object = stmnt.getObject();
 					String objStr = object.stringValue().substring(object.stringValue().lastIndexOf("#") + 1);
-					
+
 					ontPredicates.add(predStr);
 					ontTerms.add(objStr);
-					
+
 					// ontology predicate usage
 					int ontPredCount = 0;
 					if (ontPredUsage.containsKey(predStr)) {
 						ontPredCount = ontPredUsage.get(predStr);
 					}
 					ontPredUsage.put(predStr, ++ontPredCount);
-					
+
 					// ontology term usage
 					int ontTermCount = 0;
 					if (ontTermUsage.containsKey(objStr)) {
 						ontTermCount = ontTermUsage.get(objStr);
 					}
 					ontTermUsage.put(objStr, ++ontTermCount);
-					
+
 					// usedIn
 					if (predicateUsedIn.containsKey(predStr)) {
 						Set<String> usedIn = predicateUsedIn.get(predStr);
@@ -349,7 +348,7 @@ public class StatisticsResource extends BaseResource {
 						usedIn.add(entryID);
 						predicateUsedIn.put(predStr, usedIn);
 					}
-					
+
 					if (ontologyTermUsedIn.containsKey(objStr)) {
 						Set<String> usedIn = ontologyTermUsedIn.get(objStr);
 						usedIn.add(entryID);
@@ -404,6 +403,7 @@ public class StatisticsResource extends BaseResource {
 	
 	public JSONObject getKeywordStatistics(Context context) {
 		JSONObject result = new JSONObject();
+		ValueFactory vf = getRM().getValueFactory();
 		
 		URI currentUserURI = getPM().getAuthenticatedUserURI();
 		getPM().setAuthenticatedUserURI(getPM().getAdminUser().getURI());
@@ -423,42 +423,39 @@ public class StatisticsResource extends BaseResource {
 					continue;
 				}
 				
-				Set<org.openrdf.model.URI> allowedPredicates = new HashSet<org.openrdf.model.URI>();
-				allowedPredicates.add(new URIImpl(NS.dc + "subject"));
-				allowedPredicates.add(new URIImpl(NS.dcterms + "subject"));
+				Set<IRI> allowedPredicates = new HashSet<>();
+				allowedPredicates.add(vf.createIRI(NS.dc + "subject"));
+				allowedPredicates.add(vf.createIRI(NS.dcterms + "subject"));
 				
 				entryCount++;
 				Set<String> keywords = new HashSet<String>();
 				String entryID = entry.getId();
 				URI resourceURI = entry.getResourceURI();
-				Graph metadata = entry.getMetadataGraph();
-				Iterator<Statement> firstLevel = metadata.match(new URIImpl(resourceURI.toString()), null, null);
-				while (firstLevel.hasNext()) {
-					Statement stmnt = firstLevel.next();
-					org.openrdf.model.URI predicate = stmnt.getPredicate();
-					
+				Model metadata = entry.getMetadataGraph();
+				for (Statement stmnt : metadata.filter(vf.createIRI(resourceURI.toString()), null, null)) {
+					IRI predicate = stmnt.getPredicate();
+
 					if (!allowedPredicates.contains(predicate)) {
 						continue;
 					}
-					
+
 					entriesWithKeyword.add(resourceURI);
-					
+
 					Value object = stmnt.getObject();
-					
+
 					if (object instanceof Resource) {
-						Iterator<Statement> secondLevel = metadata.match((Resource) object, RDF.VALUE, null);
-						while (secondLevel.hasNext()) {
-							Value o2 = secondLevel.next().getObject();
+						for (Statement statement : metadata.filter((Resource) object, RDF.VALUE, null)) {
+							Value o2 = statement.getObject();
 							if (o2 instanceof Literal) {
 								String objStr = o2.stringValue().substring(o2.stringValue().lastIndexOf("#") + 1).toLowerCase();
 								keywords.add(objStr);
-								
+
 								int keywordCount = 0;
 								if (keywordUsage.containsKey(objStr)) {
 									keywordCount = keywordUsage.get(objStr);
 								}
 								keywordUsage.put(objStr, ++keywordCount);
-								
+
 								if (keywordUsedIn.containsKey(objStr)) {
 									Set<String> usedIn = keywordUsedIn.get(objStr);
 									usedIn.add(entryID);
@@ -473,13 +470,13 @@ public class StatisticsResource extends BaseResource {
 					} else if (object instanceof Literal) {
 						String objStr = object.stringValue().substring(object.stringValue().lastIndexOf("#") + 1).toLowerCase();
 						keywords.add(objStr);
-						
+
 						int keywordCount = 0;
 						if (keywordUsage.containsKey(objStr)) {
 							keywordCount = keywordUsage.get(objStr);
 						}
 						keywordUsage.put(objStr, ++keywordCount);
-						
+
 						if (keywordUsedIn.containsKey(objStr)) {
 							Set<String> usedIn = keywordUsedIn.get(objStr);
 							usedIn.add(entryID);
@@ -503,7 +500,7 @@ public class StatisticsResource extends BaseResource {
 			for (String key : keywords) {
 				JSONObject propStats = new JSONObject();
 				propStats.put("title", key);
-				int keywordCount = keywordUsage.get(key).intValue();
+				int keywordCount = keywordUsage.get(key);
 				int usedInCount = keywordUsedIn.get(key).size();
 				propStats.put("totalCount", keywordCount);
 				propStats.put("usedInCount", usedInCount);
@@ -567,36 +564,24 @@ public class StatisticsResource extends BaseResource {
 			//and can be found in relations
 			List<Statement> stats = user.getEntry().getRelations();
 			for (Statement s : stats){
-				org.openrdf.model.URI pred = s.getPredicate();
+				IRI pred = s.getPredicate();
 				if("http://scam.sf.net/schema#aboutPerson".equals(pred.toString())){
-					java.net.URI resourceURI = java.net.URI.create(s.getSubject().toString());
+					URI resourceURI = URI.create(s.getSubject().toString());
 					String contextId = org.entrystore.impl.Util.getContextIdFromURI(this.getRM(), resourceURI);
 					Context context = cm.getContext(contextId);
 					Set<Entry> competenceEntries = context.getByResourceURI(resourceURI);
 					Entry competenceEntry = competenceEntries.iterator().next(); //Should only be one!
-					Graph graph = competenceEntry.getMetadataGraph();
-					Iterator<Statement> compDefsStatements = graph.match( null,
-							graph.getValueFactory().createURI("http://scam.sf.net/schema#competencyDefinition"), (Resource) null);
-					
-					while (compDefsStatements.hasNext()){
-						Statement stat = compDefsStatements.next();
-						try{
-							Value compLevel  = GraphUtil.getUniqueObject(graph, stat.getSubject(), 
-									graph.getValueFactory().createURI("http://scam.sf.net/schema#competenceLevel"));
-							HashMap<String, Integer> current = CompDefToCount.get(stat.getObject().toString());
-							if (current == null){
-								current = new HashMap<String, Integer>();
-								CompDefToCount.put(stat.getObject().toString(), current);
-							}
-							Integer currentInt = current.get(compLevel.toString());
-							if(currentInt == null){
-								currentInt = new Integer(0);
-							}
-							currentInt++;
-							current.put(compLevel.toString(), currentInt);
-						} catch (GraphUtilException gue) {
-							continue;
+					Model graph = competenceEntry.getMetadataGraph();
+
+					for (Statement stat : graph.filter(null, iri("http://scam.sf.net/schema#competencyDefinition"), (Resource) null)) {
+						Value compLevel = graph.filter(stat.getSubject(), iri("http://scam.sf.net/schema#competenceLevel"), null).objects().iterator().next();
+						HashMap<String, Integer> current = CompDefToCount.computeIfAbsent(stat.getObject().toString(), k -> new HashMap<String, Integer>());
+						Integer currentInt = current.get(compLevel.toString());
+						if (currentInt == null) {
+							currentInt = 0;
 						}
+						currentInt++;
+						current.put(compLevel.toString(), currentInt);
 					}
 					break;
 				}
