@@ -16,34 +16,37 @@
 
 package org.entrystore.rest.util;
 
-import com.github.jsonldjava.sesame.SesameJSONLDParser;
-import com.github.jsonldjava.sesame.SesameJSONLDWriter;
-import info.aduna.xml.XMLReaderFactory;
+import org.eclipse.rdf4j.common.xml.XMLReaderFactory;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.Statement;
+import org.eclipse.rdf4j.model.impl.LinkedHashModel;
+import org.eclipse.rdf4j.rio.ParserConfig;
+import org.eclipse.rdf4j.rio.RDFFormat;
+import org.eclipse.rdf4j.rio.RDFHandler;
+import org.eclipse.rdf4j.rio.RDFHandlerException;
+import org.eclipse.rdf4j.rio.RDFParseException;
+import org.eclipse.rdf4j.rio.RDFParser;
+import org.eclipse.rdf4j.rio.RDFWriter;
+import org.eclipse.rdf4j.rio.helpers.BasicWriterSettings;
+import org.eclipse.rdf4j.rio.helpers.JSONLDMode;
+import org.eclipse.rdf4j.rio.helpers.JSONLDSettings;
+import org.eclipse.rdf4j.rio.helpers.StatementCollector;
+import org.eclipse.rdf4j.rio.helpers.XMLParserSettings;
+import org.eclipse.rdf4j.rio.jsonld.JSONLDParser;
+import org.eclipse.rdf4j.rio.jsonld.JSONLDWriter;
+import org.eclipse.rdf4j.rio.n3.N3ParserFactory;
+import org.eclipse.rdf4j.rio.n3.N3Writer;
+import org.eclipse.rdf4j.rio.ntriples.NTriplesParser;
+import org.eclipse.rdf4j.rio.ntriples.NTriplesWriter;
+import org.eclipse.rdf4j.rio.rdfxml.RDFXMLParser;
+import org.eclipse.rdf4j.rio.rdfxml.util.RDFXMLPrettyWriter;
+import org.eclipse.rdf4j.rio.trig.TriGParser;
+import org.eclipse.rdf4j.rio.trig.TriGWriter;
+import org.eclipse.rdf4j.rio.trix.TriXParser;
+import org.eclipse.rdf4j.rio.trix.TriXWriter;
+import org.eclipse.rdf4j.rio.turtle.TurtleParser;
+import org.eclipse.rdf4j.rio.turtle.TurtleWriter;
 import org.entrystore.repository.util.NS;
-import org.openrdf.model.Graph;
-import org.openrdf.model.Statement;
-import org.openrdf.model.impl.LinkedHashModel;
-import org.openrdf.rio.ParserConfig;
-import org.openrdf.rio.RDFFormat;
-import org.openrdf.rio.RDFHandler;
-import org.openrdf.rio.RDFHandlerException;
-import org.openrdf.rio.RDFParseException;
-import org.openrdf.rio.RDFParser;
-import org.openrdf.rio.RDFWriter;
-import org.openrdf.rio.helpers.StatementCollector;
-import org.openrdf.rio.helpers.XMLParserSettings;
-import org.openrdf.rio.n3.N3ParserFactory;
-import org.openrdf.rio.n3.N3Writer;
-import org.openrdf.rio.ntriples.NTriplesParser;
-import org.openrdf.rio.ntriples.NTriplesWriter;
-import org.openrdf.rio.rdfxml.RDFXMLParser;
-import org.openrdf.rio.rdfxml.util.RDFXMLPrettyWriter;
-import org.openrdf.rio.trig.TriGParser;
-import org.openrdf.rio.trig.TriGWriter;
-import org.openrdf.rio.trix.TriXParser;
-import org.openrdf.rio.trix.TriXWriter;
-import org.openrdf.rio.turtle.TurtleParser;
-import org.openrdf.rio.turtle.TurtleWriter;
 import org.restlet.data.MediaType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -66,9 +69,9 @@ import java.util.Map;
  */
 public class GraphUtil {
 
-	private static Logger log = LoggerFactory.getLogger(GraphUtil.class);
+	private final static Logger log = LoggerFactory.getLogger(GraphUtil.class);
 
-	private static List<MediaType> supportedMediaTypes = new ArrayList<>();
+	private final static List<MediaType> supportedMediaTypes = new ArrayList<>();
 
 	static {
 		supportedMediaTypes.add(MediaType.APPLICATION_RDF_XML);
@@ -89,7 +92,7 @@ public class GraphUtil {
 	 *               TurtleWriter
 	 * @return A String representation of the serialized Graph.
 	 */
-	public static String serializeGraph(Graph graph, Class<? extends RDFWriter> writer) {
+	public static String serializeGraph(Model graph, Class<? extends RDFWriter> writer) {
 		if (graph == null || writer == null) {
 			throw new IllegalArgumentException("Parameters must not be null");
 		}
@@ -97,8 +100,35 @@ public class GraphUtil {
 		StringWriter stringWriter = new StringWriter();
 		RDFWriter rdfWriter = null;
 		try {
+			/*
+			Constructor<? extends RDFWriter> constructor = null;
+			if (writer.equals(JSONLDWriter.class) && NS.getMap().containsKey("store")) {
+				constructor = writer.getConstructor(Writer.class, String.class);
+				rdfWriter = constructor.newInstance(stringWriter, NS.getMap().get("store"));
+			} else {
+				constructor = writer.getConstructor(Writer.class);
+				rdfWriter = constructor.newInstance(stringWriter);
+			}
+			*/
 			Constructor<? extends RDFWriter> constructor = writer.getConstructor(Writer.class);
-			rdfWriter = (RDFWriter) constructor.newInstance(stringWriter);
+			rdfWriter = constructor.newInstance(stringWriter);
+
+			if (!System.getProperties().containsKey("org.eclipse.rdf4j.rio.rdf10_plain_literals")) {
+				rdfWriter.getWriterConfig().set(BasicWriterSettings.XSD_STRING_TO_PLAIN_LITERAL, true);
+			}
+			if (!System.getProperties().containsKey("org.eclipse.rdf4j.rio.rdf10_language_literals")) {
+				rdfWriter.getWriterConfig().set(BasicWriterSettings.RDF_LANGSTRING_TO_LANG_LITERAL, true);
+			}
+			if (!System.getProperties().containsKey("org.eclipse.rdf4j.rio.jsonld.optimize")) {
+				rdfWriter.getWriterConfig().set(JSONLDSettings.OPTIMIZE, true);
+			}
+			if (!System.getProperties().containsKey("org.eclipse.rdf4j.rio.jsonld.use_native_types")) {
+				rdfWriter.getWriterConfig().set(JSONLDSettings.USE_NATIVE_TYPES, true);
+			}
+			if (!System.getProperties().containsKey("org.eclipse.rdf4j.rio.jsonld.hierarchical_view")) {
+				rdfWriter.getWriterConfig().set(JSONLDSettings.HIERARCHICAL_VIEW, true);
+			}
+			rdfWriter.getWriterConfig().set(JSONLDSettings.JSONLD_MODE, JSONLDMode.COMPACT);
 		} catch (Exception e) {
 			log.error(e.getMessage());
 		}
@@ -108,11 +138,16 @@ public class GraphUtil {
 		}
 
 		try {
+			rdfWriter.startRDF();
 			Map<String, String> namespaces = NS.getMap();
 			for (String nsName : namespaces.keySet()) {
+				/*
+				if ("store".equals(nsName)) {
+					continue;
+				}
+				*/
 				rdfWriter.handleNamespace(nsName, namespaces.get(nsName));
 			}
-			rdfWriter.startRDF();
 			for (Statement statement : graph) {
 				rdfWriter.handleStatement(statement);
 			}
@@ -123,16 +158,16 @@ public class GraphUtil {
 		return stringWriter.toString();
 	}
 
-	public static void serializeGraph(Graph graph, RDFWriter rdfWriter) {
+	public static void serializeGraph(Model graph, RDFWriter rdfWriter) {
 		if (graph == null || rdfWriter == null) {
 			throw new IllegalArgumentException("Parameters must not be null");
 		}
 		try {
+			rdfWriter.startRDF();
 			Map<String, String> namespaces = NS.getMap();
 			for (String nsName : namespaces.keySet()) {
 				rdfWriter.handleNamespace(nsName, namespaces.get(nsName));
 			}
-			rdfWriter.startRDF();
 			for (Statement statement : graph) {
 				rdfWriter.handleStatement(statement);
 			}
@@ -148,7 +183,7 @@ public class GraphUtil {
 	 *                        RDFXMLParser, TriGParser, TriXParser, TurtleParser
 	 * @return A String representation of the serialized Graph.
 	 */
-	public static Graph deserializeGraph(String serializedGraph, RDFParser parser) {
+	public static Model deserializeGraph(String serializedGraph, RDFParser parser) {
 		try {
 			return deserializeGraphUnsafe(serializedGraph, parser);
 		} catch (RDFHandlerException | RDFParseException | IOException e) {
@@ -157,7 +192,7 @@ public class GraphUtil {
 		}
 	}
 
-	public static Graph deserializeGraphUnsafe(String serializedGraph, RDFParser parser) throws RDFParseException, RDFHandlerException, IOException {
+	public static Model deserializeGraphUnsafe(String serializedGraph, RDFParser parser) throws RDFParseException, RDFHandlerException, IOException {
 		if (serializedGraph == null || parser == null) {
 			throw new IllegalArgumentException("Parameters must not be null");
 		}
@@ -170,7 +205,7 @@ public class GraphUtil {
 		return new LinkedHashModel(collector.getStatements());
 	}
 
-	public static Graph deserializeGraph(String graphString, MediaType mediaType) {
+	public static Model deserializeGraph(String graphString, MediaType mediaType) {
 		try {
 			return deserializeGraphUnsafe(graphString, mediaType);
 		} catch (RDFHandlerException | RDFParseException | IOException e) {
@@ -179,8 +214,8 @@ public class GraphUtil {
 		return null;
 	}
 
-	public static Graph deserializeGraphUnsafe(String graphString, MediaType mediaType) throws RDFHandlerException, IOException, RDFParseException {
-		Graph deserializedGraph = null;
+	public static Model deserializeGraphUnsafe(String graphString, MediaType mediaType) throws RDFHandlerException, IOException, RDFParseException {
+		Model deserializedGraph = null;
 		if (mediaType.equals(MediaType.APPLICATION_JSON) || mediaType.getName().equals("application/rdf+json")) {
 			deserializedGraph = RDFJSON.rdfJsonToGraph(graphString);
 		} else if (mediaType.equals(MediaType.APPLICATION_RDF_XML)) {
@@ -200,12 +235,12 @@ public class GraphUtil {
 		} else if (mediaType.getName().equals(RDFFormat.TRIG.getDefaultMIMEType())) {
 			deserializedGraph = deserializeGraphUnsafe(graphString, new TriGParser());
 		} else if (mediaType.getName().equals(RDFFormat.JSONLD.getDefaultMIMEType())) {
-			deserializedGraph = deserializeGraphUnsafe(graphString, new SesameJSONLDParser());
+			deserializedGraph = deserializeGraphUnsafe(graphString, new JSONLDParser());
 		}
 		return deserializedGraph;
 	}
 
-	public static String serializeGraph(Graph graph, MediaType mediaType) {
+	public static String serializeGraph(Model graph, MediaType mediaType) {
 		String serializedGraph = null;
 		if (mediaType.equals(MediaType.APPLICATION_JSON) || mediaType.getName().equals("application/rdf+json")) {
 			serializedGraph = RDFJSON.graphToRdfJson(graph);
@@ -222,7 +257,7 @@ public class GraphUtil {
 		} else if (mediaType.getName().equals(RDFFormat.TRIG.getDefaultMIMEType())) {
 			serializedGraph = serializeGraph(graph, TriGWriter.class);
 		} else if (mediaType.getName().equals(RDFFormat.JSONLD.getDefaultMIMEType())) {
-			serializedGraph = serializeGraph(graph, SesameJSONLDWriter.class);
+			serializedGraph = serializeGraph(graph, JSONLDWriter.class);
 		} else {
 			// fallback
 			serializedGraph = serializeGraph(graph, TurtleWriter.class);
@@ -257,7 +292,7 @@ public class GraphUtil {
 		parser.setParserConfig(constructSafeXmlParserConfig());
 		if (mediaType.equals(MediaType.APPLICATION_JSON) || mediaType.getName().equals("application/rdf+json")) {
 			// we have special treatment of RDF/JSON here because it does not implement the Parser interface
-			Graph g = RDFJSON.rdfJsonToGraph(rdf);
+			Model g = RDFJSON.rdfJsonToGraph(rdf);
 			if (g != null) {
 				return "There was an error parsing the RDF/JSON payload";
 			} else {
@@ -275,7 +310,7 @@ public class GraphUtil {
 		} else if (mediaType.getName().equals(RDFFormat.TRIG.getDefaultMIMEType())) {
 			parser = new TriGParser();
 		} else if (mediaType.getName().equals(RDFFormat.JSONLD.getDefaultMIMEType())) {
-			parser = new SesameJSONLDParser();
+			parser = new JSONLDParser();
 		}
 
 		String error = null;
